LayoutTests/imported/w3c/web-platform-tests/service-workers/service-worker/unregister-controller.https-expected.txt



FAIL Unregister does not affect existing controller assert_unreached: unexpected rejection: assert_equals: controller should intercept requests expected "intercepted by service worker" but got "" Reached unreachable code
FAIL Unregister prevents control of subsequent navigations assert_unreached: unexpected rejection: assert_equals: requests should not be intercepted expected "a simple text file\n" but got "" Reached unreachable code
PASS Unregister prevents new controllee even if registration is still in use 
